break evaluate function
