+++ src/evaluator/evaluator.go
@@ -96,137 +96,6 @@ func Eval(node ast.Node, env *object.Environment) object.Object {
 	return nil
 }
 
-func evalProgram(program *ast.Program, env *object.Environment) object.Object {
-	var result object.Object
-	for _, statement := range program.Statements {
-		result = Eval(statement, env)
-		switch result := result.(type) {
-		case *object.ReturnValue:
-			return result.Value
-		case *object.Error:
-			return result
-		}
-	}
-	return result
-}
-
-func evalBlockStatement(block *ast.BlockStatement, env *object.Environment) object.Object {
-	var result object.Object
-	for _, statement := range block.Statements {
-		result = Eval(statement, env)
-		if result != nil {
-			rt := result.Type()
-			if rt == object.RETURN_VALUE_OBJ || rt == object.ERROR_OBJ {
-				return result
-			}
-		}
-	}
-	return result
-}
-
-func nativeBoolToBooleanObject(input bool) *object.Boolean {
-	if input {
-		return TRUE
-	}
-	return FALSE
-}
-
-func evalPrefixExpression(operator string, right object.Object) object.Object {
-	switch operator {
-	case "!":
-		return evalBangOperatorExpression(right)
-	case "-":
-		return evalMinusPrefixOperatorExpression(right)
-	default:
-		return newError("unknown operator: %s%s", operator, right.Type())
-	}
-}
-
-func evalBangOperatorExpression(right object.Object) object.Object {
-	switch right {
-	case TRUE:
-		return FALSE
-	case FALSE:
-		return TRUE
-	case NULL:
-		return TRUE
-	default:
-		return FALSE
-	}
-}
-
-func evalMinusPrefixOperatorExpression(right object.Object) object.Object {
-	if right.Type() != object.INTEGER_OBJ {
-		return newError("unknown operator:-%s", right.Type())
-	}
-	value := right.(*object.Integer).Value.Neg()
-	return &object.Integer{Value: value}
-}
-
-func evalInfixExpression(
-	operator string,
-	left, right object.Object,
-) object.Object {
-	switch {
-	case left.Type() == object.INTEGER_OBJ && right.Type() == object.INTEGER_OBJ:
-		return evalIntegerInfixExpression(operator, left, right)
-	case left.Type() == object.STRING_OBJ && right.Type() == object.STRING_OBJ:
-		return evalStringInfixExpression(operator, left, right)
-	case operator == "==":
-		return nativeBoolToBooleanObject(left == right)
-	case operator == "!=":
-		return nativeBoolToBooleanObject(left != right)
-	case left.Type() != right.Type():
-		return newError("type mismatch: %s %s %s",
-			left.Type(), operator, right.Type())
-	default:
-		return newError("unknown operator: %s %s %s",
-			left.Type(), operator, right.Type())
-	}
-}
-
-func evalIntegerInfixExpression(
-	operator string,
-	left, right object.Object,
-) object.Object {
-	leftVal := left.(*object.Integer).Value
-	rightVal := right.(*object.Integer).Value
-	switch operator {
-	case "+":
-		return &object.Integer{Value: leftVal.Add(rightVal)}
-	case "-":
-		return &object.Integer{Value: leftVal.Sub(rightVal)}
-	case "*":
-		return &object.Integer{Value: leftVal.Mul(rightVal)}
-	case "/":
-		return &object.Integer{Value: leftVal.Div(rightVal)}
-	case "<":
-		return nativeBoolToBooleanObject(leftVal.LessThan(rightVal))
-	case ">":
-		return nativeBoolToBooleanObject(leftVal.GreaterThan(rightVal))
-	case "==":
-		return nativeBoolToBooleanObject(leftVal.Equal(rightVal))
-	case "!=":
-		return nativeBoolToBooleanObject(!leftVal.Equal(rightVal))
-	default:
-		return newError("unknown operator: %s %s %s",
-			left.Type(), operator, right.Type())
-	}
-}
-
-func evalIfExpression(ie *ast.IfExpression, env *object.Environment) object.Object {
-	condition := Eval(ie.Condition, env)
-	if isError(condition) {
-		return condition
-	}
-	if isTruthy(condition) {
-		return Eval(ie.Consequence, env)
-	} else if ie.Alternative != nil {
-		return Eval(ie.Alternative, env)
-	} else {
-		return NULL
-	}
-}
 func isTruthy(obj object.Object) bool {
 	switch obj {
 	case NULL:
@@ -249,94 +118,4 @@ func isError(obj object.Object) bool {
 		return obj.Type() == object.ERROR_OBJ
 	}
 	return false
-}
-
-func evalIdentifier(
-	node *ast.Identifier,
-	env *object.Environment,
-) object.Object {
-	val, ok := env.Get(node.Value)
-	if !ok {
-		return newError("identifier not found: " + node.Value)
-	}
-	return val
-}
-
-func evalExpressions(
-	exps []ast.Expression,
-	env *object.Environment,
-) []object.Object {
-	var result []object.Object
-	for _, e := range exps {
-		evaluated := Eval(e, env)
-		if isError(evaluated) {
-			return []object.Object{evaluated}
-		}
-		result = append(result, evaluated)
-	}
-	return result
-}
-
-func applyFunction(fn object.Object, args []object.Object) object.Object {
-	function, ok := fn.(*object.Function)
-	if !ok {
-		return newError("not a function: %s", fn.Type())
-	}
-	extendedEnv := extendFunctionEnv(function, args)
-	evaluated := Eval(function.Body, extendedEnv)
-	return unwrapReturnValue(evaluated)
-}
-
-func extendFunctionEnv(
-	fn *object.Function,
-	args []object.Object,
-) *object.Environment {
-	env := object.NewEnclosedEnvironment(fn.Env)
-	for paramIdx, param := range fn.Parameters {
-		env.Set(param.Value, args[paramIdx])
-	}
-	return env
-}
-
-func unwrapReturnValue(obj object.Object) object.Object {
-	if returnValue, ok := obj.(*object.ReturnValue); ok {
-		return returnValue.Value
-	}
-	return obj
-}
-
-func evalStringInfixExpression(
-	operator string,
-	left, right object.Object,
-) object.Object {
-	leftVal := left.(*object.String).Value
-	rightVal := right.(*object.String).Value
-	switch operator {
-	case "+":
-		return &object.String{Value: leftVal + rightVal}
-	case "==":
-		return nativeBoolToBooleanObject(leftVal == rightVal)
-	default:
-		return newError("unknown operator: %s %s %s",
-			left.Type(), operator, right.Type())
-	}
-}
-
-func evalIndexExpression(left, index object.Object) object.Object {
-	switch {
-	case left.Type() == object.ARRAY_OBJ && index.Type() == object.INTEGER_OBJ:
-		return evalArrayIndexExpression(left, index)
-	default:
-		return newError("index operator not supported: %s", left.Type())
-	}
-}
-
-func evalArrayIndexExpression(array, index object.Object) object.Object {
-	arrayObject := array.(*object.Array)
-	idx := index.(*object.Integer).Value.IntPart()
-	maxValue := int64(len(arrayObject.Elements) - 1)
-	if idx < 0 || idx > maxValue {
-		return NULL
-	}
-	return arrayObject.Elements[idx]
 }
