+++ src/evaluator/return.go
@@ -1,0 +1,10 @@
+package evaluator
+
+import "github.com/sevenreup/chewa/src/object"
+
+func unwrapReturnValue(obj object.Object) object.Object {
+	if returnValue, ok := obj.(*object.ReturnValue); ok {
+		return returnValue.Value
+	}
+	return obj
+}
